# PERF: improve find_common_type perf for special case of all-equal dtypes