@jbrockmendel 2021-11-23
nice speedup!